add datetime filtering and make it more pythonic
